created input style, added sign up button and image
