+++ src/app/globals.css
@@ -110,6 +110,16 @@
   --sidebar-ring: oklch(0.556 0 0);
 }
 
+.inputStyle {
+  color: #fff;         /* biały tekst */
+  opacity: 1;          /* pełna widoczność */
+  background: rgba(125, 125, 125, 0.5);
+  border-radius: 24px;
+  padding-inline: 20px; 
+  padding-top: 6px;
+  padding-bottom: 6px;
+}
+
 @layer base {
   * {
     @apply border-border outline-ring/50;
@@ -119,6 +129,3 @@
   }
 }
 
-.buttonStyle{
-  @apply flex items-center justify-center w-full;
-}
