+++ src/app/page.tsx
@@ -1,5 +1,4 @@
-//import Image from "next/image";
-import SignInForm from "./LoginForm";
+import SignInForm from "./SignInForm";
 
 export default function Page() {
   return (
